Fixed a few javascript issues and implemented team matching scores and skill endorsement.
Users can now view team members when joining a hackathon and see if they are a match or not from filtering the users skills and comparing them to the user and giving a percentage based result. They can also connect with those and message them who have joined the same hackathon as them. Also on the projects side of the website there has been changes to how commit messages are displayed and also added a new feature where every users skills that they have put on their profile are displayed on the team members section and other users can endorse and interact to show the user is talented in whichever skill they have. Also fixed a few minor issues with missing javascript files for the resource page and now they are working fine as well.